+++ src/interfaces/exceptions/error-response.ts
@@ -1,4 +1,4 @@
 import type { AxiosError } from 'axios';
 import { Response } from './../http/response';
 
-export type ErrorRepsonse = AxiosError<Response, any>;
+export type ErrorResponse = AxiosError<Response, any>;